+++ src/features/user-stats/get-data.ts
@@ -3,27 +3,55 @@ import { User } from "firebase/auth"
 import { collection, getDocs, orderBy, query, where } from "firebase/firestore"
 import { UserSerie } from "react-charts"
 import { DayData, PokoyChartData, UserStatsData } from "shared/types"
-import { roundToFirstDecimalPlace } from "shared/utils/roundToSecondDecimalPlace"
-import {
-  MINS_IN_HOUR,
-  SECONDARY_AXIS_LABEL,
-  SECS_IN_DAY,
-  TERTIARY_AXIS_LABEL,
-} from "./constants"
+import { THIRD_PART } from "./constants"
 import { getFullRange } from "./get-full-range"
+import {
+  getAverageMeditationPerDay,
+  getPseudoDayData,
+  sliceDaysDataRange,
+  transformDayDataToChartData,
+} from "./utils"
 
-const INITIAL_MEDITATION_DURATION = 0
-
-// TODO: refactor this module
-export const fetchAndsetChartData = async (
+export const fetchAndSetChartData = async (
   setDataToComponentState: (data: UserSerie<PokoyChartData>[]) => void,
   user: User
 ): Promise<void> => {
   const daysWithMeditations = await fetchDays(user)
-  const dayDataFullRange = getFullRange(daysWithMeditations)
-  const chartData = transformDayDataToChartData(dayDataFullRange)
+  const daysDataFullRange = getFullRange(daysWithMeditations)
+
+  const additionalDataLength = Math.round(daysDataFullRange.length * THIRD_PART)
+  const additionalDaysData = await getForesightChartData(
+    sliceDaysDataRange(daysDataFullRange),
+    user,
+    additionalDataLength
+  )
+  const daysDataWithForesight = daysDataFullRange.concat(additionalDaysData)
+  const chartData = transformDayDataToChartData(daysDataWithForesight)
 
   return setDataToComponentState(chartData)
+}
+
+// TODO: refactor this method
+// eslint-disable-next-line max-statements
+export async function getForesightChartData(
+  daysData: DayData[],
+  user: User,
+  additionalDataLength: number
+) {
+  const lastData = daysData[daysData.length - 1]
+  const lastDateTimestampSeconds = lastData.timestamp.seconds
+
+  const stats = await fetchStats(user)
+  const averageMeditationDuration = getAverageMeditationPerDay(stats)
+  const daysToNextMilestone = additionalDataLength
+
+  const additionalDaysData: DayData[] = new Array(daysToNextMilestone)
+    .fill(null)
+    .map((_, i) =>
+      getPseudoDayData(i, lastDateTimestampSeconds, averageMeditationDuration)
+    )
+
+  return additionalDaysData
 }
 
 export const getStats = async (
@@ -40,27 +68,6 @@ export const getStats = async (
   setDataToComponentState(statsData)
 }
 
-export const getTotalInHours = (minutes: number): number => {
-  return Math.floor(minutes / MINS_IN_HOUR)
-}
-
-export const getAverage = (statsData: UserStatsData) => {
-  if (!statsData || !statsData.firstMeditationDate) {
-    return null
-  }
-
-  const { firstMeditationDate } = statsData
-  const statsMillisecondsDiff =
-    Date.now() - firstMeditationDate.toDate().getTime()
-  const statsRangeInDays = statsMillisecondsDiff / SECS_IN_DAY
-
-  const average = roundToFirstDecimalPlace(
-    statsData.totalDuration / statsRangeInDays
-  )
-
-  return average
-}
-
 async function fetchStats(user: User): Promise<UserStatsData> {
   const statsColRef = collection(firestore, "stats")
   const statsQuery = query(statsColRef, where("userId", "==", user.uid))
@@ -68,57 +75,6 @@ async function fetchStats(user: User): Promise<UserStatsData> {
   const statsData = daysColSnapshot?.docs[0]?.data() as UserStatsData
 
   return statsData
-}
-
-function transformDayDataToChartData(
-  dayDataFullRange: DayData[]
-): UserSerie<PokoyChartData>[] {
-  const daysWithMeditationsAsAxis: PokoyChartData[] = dayDataFullRange.map(
-    (d) => ({
-      primary: new Date(d.timestamp.toDate().toDateString()),
-      secondary: d.totalDuration,
-    })
-  )
-  const totalDurationsAxisData: PokoyChartData[] = getTotalDurationsAsAxisData(
-    daysWithMeditationsAsAxis
-  )
-
-  const secondaryAxisData = {
-    label: SECONDARY_AXIS_LABEL,
-    data: daysWithMeditationsAsAxis,
-    id: "2",
-    secondaryAxisId: "2",
-  }
-  const tertiaryAxisData = {
-    label: TERTIARY_AXIS_LABEL,
-    data: totalDurationsAxisData,
-    id: "1",
-    // secondaryAxisId: "1"
-  }
-
-  const chartData = [secondaryAxisData, tertiaryAxisData]
-  return chartData
-}
-
-function getTotalDurationsAsAxisData(
-  daysWithMeditationsAsAxis: PokoyChartData[]
-): PokoyChartData[] {
-  const totalDurationAsAxisData = daysWithMeditationsAsAxis.reduce(
-    (acc, d, i) => {
-      const prevTotal = acc[i - 1]?.secondary || INITIAL_MEDITATION_DURATION
-      const total = roundToFirstDecimalPlace(d.secondary / 60 + prevTotal)
-      return [
-        ...acc,
-        {
-          primary: d.primary,
-          secondary: total,
-        },
-      ]
-    },
-    [] as PokoyChartData[]
-  )
-
-  return totalDurationAsAxisData
 }
 
 async function fetchDays(user: User): Promise<DayData[]> {